requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
albucore=0.0.23=pypi_0
albumentations=2.0.5=pypi_0
altair=5.5.0=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.9.0=pypi_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.3.0=pypi_0
asgiref=3.8.1=pypi_0
astor=0.8.1=pypi_0
asttokens=3.0.0=pypi_0
async-lru=2.0.5=pypi_0
attrs=25.3.0=pypi_0
babel=2.17.0=pypi_0
beautifulsoup4=4.13.4=pypi_0
bleach=6.2.0=pypi_0
blinker=1.9.0=pypi_0
bzip2=1.0.8=h2466b09_7
ca-certificates=2025.2.25=haa95532_0
cachetools=5.5.2=pypi_0
cairo=1.18.4=h5782bbf_0
camelot-py=1.0.0=pypi_0
certifi=2025.1.31=pypi_0
cffi=1.17.1=pypi_0
chardet=5.2.0=pypi_0
charset-normalizer=3.4.1=pypi_0
click=8.1.8=pypi_0
colorama=0.4.6=pypi_0
comm=0.2.2=pypi_0
contourpy=1.3.0=pypi_0
cryptography=44.0.2=pypi_0
cycler=0.12.1=pypi_0
cython=3.0.12=pypi_0
debugpy=1.8.14=pypi_0
decorator=5.2.1=pypi_0
defusedxml=0.7.1=pypi_0
diskcache=5.6.3=pypi_0
django=4.2.20=pypi_0
et-xmlfile=2.0.0=pypi_0
eval-type-backport=0.2.2=pypi_0
exceptiongroup=1.2.2=pypi_0
executing=2.2.0=pypi_0
faker=37.1.0=pypi_0
fastjsonschema=2.21.1=pypi_0
filelock=3.18.0=pypi_0
fire=0.7.0=pypi_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=h77eed37_3
fontconfig=2.15.0=h765892d_1
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.57.0=pypi_0
fqdn=1.5.1=pypi_0
freetype=2.13.3=h0b5ce68_0
fsspec=2025.3.2=pypi_0
gitdb=4.0.12=pypi_0
gitpython=3.1.44=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.8=pypi_0
httpx=0.28.1=pypi_0
huggingface-hub=0.30.2=pypi_0
icu=75.1=he0c23c2_0
idna=3.10=pypi_0
imageio=2.37.0=pypi_0
importlib-metadata=8.6.1=pypi_0
importlib-resources=6.5.2=pypi_0
ipykernel=6.29.5=pypi_0
ipython=8.18.1=pypi_0
ipywidgets=8.1.6=pypi_0
isoduration=20.11.0=pypi_0
jedi=0.19.2=pypi_0
jinja2=3.1.6=pypi_0
json5=0.12.0=pypi_0
jsonpointer=3.0.0=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2025.4.1=pypi_0
jupyter=1.1.1=pypi_0
jupyter-client=8.6.3=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter-core=5.7.2=pypi_0
jupyter-events=0.12.0=pypi_0
jupyter-lsp=2.2.5=pypi_0
jupyter-server=2.15.0=pypi_0
jupyter-server-terminals=0.5.3=pypi_0
jupyterlab=4.4.1=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
jupyterlab-server=2.27.3=pypi_0
jupyterlab-widgets=3.0.14=pypi_0
kiwisolver=1.4.7=pypi_0
krb5=1.21.3=hdf4eb48_0
lazy-loader=0.4=pypi_0
lcms2=2.17=hbcf6048_0
lerc=4.0.0=h6470a55_1
libcurl=8.13.0=h88aaa65_0
libdeflate=1.23=h76ddb4d_0
libexpat=2.7.0=he0c23c2_0
libffi=3.4.6=h537db12_1
libglib=2.84.1=h7025463_0
libiconv=1.18=h135ad9c_1
libintl=0.22.5=h5728263_3
libjpeg-turbo=3.1.0=h2466b09_0
liblzma=5.8.1=h2466b09_0
libpng=1.6.47=had7236b_0
libssh2=1.11.1=he619c9f_0
libtiff=4.7.0=h797046b_3
libzlib=1.3.1=h2466b09_2
llama-cpp-python=0.2.30=pypi_0
lmdb=1.6.2=pypi_0
lxml=5.3.2=pypi_0
markupsafe=3.0.2=pypi_0
matplotlib=3.9.4=pypi_0
matplotlib-inline=0.1.7=pypi_0
mistune=3.1.3=pypi_0
mpmath=1.3.0=pypi_0
narwhals=1.37.0=pypi_0
nbclient=0.10.2=pypi_0
nbconvert=7.16.6=pypi_0
nbformat=5.10.4=pypi_0
nest-asyncio=1.6.0=pypi_0
networkx=3.2.1=pypi_0
notebook=7.4.1=pypi_0
notebook-shim=0.2.4=pypi_0
numpy=2.0.2=pypi_0
opencv-contrib-python=4.11.0.86=pypi_0
opencv-python=4.11.0.86=pypi_0
opencv-python-headless=4.11.0.86=pypi_0
openjpeg=2.5.3=h4d64b90_0
openpyxl=3.1.5=pypi_0
openssl=3.5.0=ha4e3fda_0
opt-einsum=3.3.0=pypi_0
overrides=7.7.0=pypi_0
packaging=24.2=pypi_0
paddle-bfloat=0.1.7=pypi_0
paddleocr=2.10.0=pypi_0
paddlepaddle=2.5.0=pypi_0
pandas=2.2.3=pypi_0
pandocfilters=1.5.1=pypi_0
parso=0.8.4=pypi_0
pcre2=10.44=h3d7b363_2
pdf2image=1.17.0=pypi_0
pdfminer-six=20250327=pypi_0
pdfplumber=0.11.6=pypi_0
pillow=11.2.1=pypi_0
pip=25.0=py39haa95532_0
pixman=0.44.2=had0cd8c_0
platformdirs=4.3.7=pypi_0
poppler=25.02.0=h82ab39c_2
poppler-data=0.4.12=hd8ed1ab_0
prometheus-client=0.21.1=pypi_0
prompt-toolkit=3.0.51=pypi_0
protobuf=3.20.2=pypi_0
psutil=7.0.0=pypi_0
pure-eval=0.2.3=pypi_0
pyarrow=20.0.0=pypi_0
pyclipper=1.3.0.post6=pypi_0
pycparser=2.22=pypi_0
pydantic=2.11.3=pypi_0
pydantic-core=2.33.1=pypi_0
pydeck=0.9.1=pypi_0
pygments=2.19.1=pypi_0
pyparsing=3.2.3=pypi_0
pypdf=3.17.4=pypi_0
pypdf2=3.0.1=pypi_0
pypdfium2=4.30.1=pypi_0
pytesseract=0.3.13=pypi_0
python=3.9.21=h8205438_1
python-dateutil=2.9.0.post0=pypi_0
python-docx=1.1.2=pypi_0
python-json-logger=3.3.0=pypi_0
pytz=2025.2=pypi_0
pywin32=310=pypi_0
pywinpty=2.0.15=pypi_0
pyyaml=6.0.2=pypi_0
pyzmq=26.4.0=pypi_0
qrcode=8.1=pypi_0
rapidfuzz=3.13.0=pypi_0
referencing=0.36.2=pypi_0
regex=2024.11.6=pypi_0
reportlab=4.4.0=pypi_0
requests=2.32.3=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rpds-py=0.24.0=pypi_0
safetensors=0.5.3=pypi_0
scikit-image=0.24.0=pypi_0
scipy=1.13.1=pypi_0
send2trash=1.8.3=pypi_0
setuptools=75.8.0=py39haa95532_0
shapely=2.0.7=pypi_0
simsimd=6.2.1=pypi_0
six=1.17.0=pypi_0
smmap=5.0.2=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.7=pypi_0
sqlite=3.45.3=h2bbff1b_0
sqlparse=0.5.3=pypi_0
stack-data=0.6.3=pypi_0
streamlit=1.44.1=pypi_0
stringzilla=3.12.5=pypi_0
sympy=1.13.3=pypi_0
tabulate=0.9.0=pypi_0
tenacity=9.1.2=pypi_0
termcolor=3.0.1=pypi_0
terminado=0.18.1=pypi_0
tifffile=2024.8.30=pypi_0
tinycss2=1.4.0=pypi_0
tokenizers=0.21.1=pypi_0
toml=0.10.2=pypi_0
tomli=2.2.1=pypi_0
torch=2.7.0=pypi_0
tornado=6.4.2=pypi_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pypi_0
transformers=4.51.3=pypi_0
types-python-dateutil=2.9.0.20241206=pypi_0
typing-extensions=4.13.2=pypi_0
typing-inspection=0.4.0=pypi_0
tzdata=2025.2=pypi_0
ucrt=10.0.22621.0=h57928b3_1
uri-template=1.3.0=pypi_0
urllib3=2.4.0=pypi_0
vc=14.42=haa95532_5
vc14_runtime=14.42.34438=hfd919c2_26
vs2015_runtime=14.42.34438=h7142326_26
watchdog=6.0.0=pypi_0
wcwidth=0.2.13=pypi_0
webcolors=24.11.1=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.8.0=pypi_0
wheel=0.45.1=py39haa95532_0
widgetsnbextension=4.0.14=pypi_0
zipp=3.21.0=pypi_0
zstd=1.5.7=hbeecb71_2
